Use komodo-testing for testing PRs

+++ ci/jenkins/komodo.sh
@@ -13,11 +13,15 @@ GCC_VERSION=7.3.0 CMAKE_VERSION=3.10.2 source ${SDPSOFT}/env.sh
 
 GIT=${SDPSOFT}/bin/git
 
-EV=$(cat ${RELEASE_PATH}/${RELEASE_NAME} | grep "${PROJECT}:" -A2 | grep "version:")
-EV=($EV)    # split the string "version: vX.X.X"
-EV=${EV[1]} # extract the version
-echo "Using ${PROJECT} version ${EV}"
-$GIT checkout $EV
+if [[ -z "${sha1// }" ]]; then
+    # this is not a PR build, the komodo everest verison is checked out
+    EV=$(cat ${RELEASE_PATH}/${RELEASE_NAME} | grep "${PROJECT}:" -A2 | grep "version:")
+    EV=($EV)    # split the string "version: vX.X.X"
+    EV=${EV[1]} # extract the version
+    EV=${EV%"+py3"}
+    echo "Using ${PROJECT} version ${EV}"
+    $GIT checkout $EV
+fi
 
 rm -rf build
 mkdir build
@@ -32,7 +36,9 @@ cmake .. -DBUILD_TESTS=ON \
 -DCMAKE_CXX_FLAGS='-Wno-unused-result'
 make -j 12
 #removing built libs in order to ensure we are using libs from komodo
-rm -r lib64
+if [[ -z "${sha1// }" ]]; then
+    rm -r lib64
+fi
 
 echo "running ctest"
 ctest --output-on-failure
